requirements.txt
click
types-setuptools
python-dotenv
PyYAML